feat: Add empty inventory message when there are no items
This commit adds a new feature to the inventory management screen. When there are no items in the inventory, an empty inventory message is displayed. This message informs the user that there are currently no items and provides instructions on how to add a new item. The message is centered on the screen and styled according to the app's design guidelines.