feat: Add AddItemDialog to InventoryManagementScreen
This commit adds the `AddItemDialog` component to the `InventoryManagementScreen` in order to allow users to add new items to the inventory. The dialog is triggered by clicking the floating action button and provides a form for entering the details of the new item. This feature enhances the functionality of the inventory management screen and improves the user experience.