feat: Update onDeleteClick in InventoryManagementScreen
This commit updates the `onDeleteClick` function in the `InventoryManagementScreen` to handle the event of deleting an item. The previous implementation had a placeholder function, but now it calls the `onEvent` function in the `InventoryManagementScreenModel` with the `Event.DeleteItem` event, passing the item to be deleted as a parameter. This enhancement improves the functionality of the inventory management screen by enabling the deletion of items.