feat: Add increment quantity functionality to InventoryManagementScreen
The code changes introduce the ability to increment the quantity of an inventory item in the InventoryManagementScreen. This enhancement allows users to easily update the stock quantity of an item. It includes adding a new event, DisplayIncrementQuantityDialog, to display a dialog for entering the additional quantity. The incrementQuanitityInventoryItem function has been implemented to update the stock quantity of the selected item. These changes improve the usability and flexibility of the inventory management screen.