feat: Update UserProfileScreenModel to observe field changes
This commit updates the `UserProfileScreenModel` file in the `feature_profile` package. It introduces a new method `observeFieldChanges()` that uses the `combine` function from `kotlinx.coroutines.flow` to observe changes in the `fullNameField` and `phoneNumberField` fields. When either of these fields change, the `UIState` is updated with the new values. This improves the functionality of the user profile screen by keeping the UI in sync with the user's input.